[eos] Add PopulationMonteCarloSampler
The addition of the PMC sampler mandates usage of
version 1.2 of the external library libpmc, obtainable
from

  http://www2.iap.fr/users/kilbinge/CosmoPMC/

After installation to e.g. /opt/pkgs/pmclib-1.2/ you
need to fix the header files as follows:

  find /opt/pkgs/pmclib-1.2/include -name "*.h" \
    -exec sed -i \
    -e 's/#include "errorlist\.h"/#include <pmctools\/errorlist.h>/' \
    -e 's/#include "io\.h"/#include <pmctools\/io.h>/' \
    -e 's/#include "mvdens\.h"/#include <pmctools\/mvdens.h>/' \
    -e 's/#include "maths\.h"/#include <pmctools\/maths.h>/' \
    -e 's/#include "maths_base\.h"/#include <pmctools\/maths_base.h>/' \
    {} \;

  sed -i \
    -e 's/^double fmin(double/\/\/&/' \
    -e 's/^double fmax(double/\/\/&/' \
    /opt/pkgs/pmclib-1.2/include/pmctools/maths.h

In order to build EOS with PMC support, you need to add the option

  --with-pmc=/opt/pkgs/pmclib-1.2

to your configure command. Not specifying the location of libpmc, or adding

  --with-pmc=no

to your configure command line will disable the PMC support in EOS.