[travis] Fix builds on MacOSX
So far, we had to manually remove /usr/local/include/c++ in order for the TravisCI build to succeed on MacOSX.
Apparently this is not only not needed anymore; the directory is not even present anymore. Hence, the rm command
fails, and leads to an error. Removing the command from 'before_install'.