[rare-b-decays] Fix bug in handling the tensor form factors for Lambda_b -> Lambda ll
When converting from the old code, which estimated ratios of tensor over vector form factors via
the heavy quark expansion, to the new code, a bug was introduced when using the full form factor
information.

To avoid this bug in the future, the obsolete old code using form factor ratios has been removed
entirely.

Thanks to Thomas Blake and Stefan Meinel for their help in debugging the code.