[form-factors] Fix calculation of the B-meson mass from the B->pi sum rules
While the previous approach to determine the B-meson mass using a numerical
derivative w.r.t -1/M^2 worked for the F correlator (i.e., for the f_+ form
factor), it fails to work for Ftilde (i.e., for the f_0 form factor). To have
a consistent approach to the determination of the B-meson mass, we calculate
it now across all form factors by selecting an appropriate weight function
for the integrand.