[form-factors] Fix bug in implementation of B->pi::f_T@DKMMO2008
In the implementation of the twist-3 contribution to the correlator FT,
which affects the f_T form factor, a typo appears in the fourth line of
eq. (B.35): a plus distribution does not cover all the rho-dependent terms
in the expression, but it finite without applying the plus distribution
correctly. The bug fix aligns our implementaiton with the numerical
implementation used in [DKMMO:2008A].