[b-decays] Fix sign errors in bto3lnu observables
Fixes sign errors in terms proportional to f_B for the quintuple
differential decay width and the asymmetry, since these terms are
proportional to the charge of the B-meson which lead to the sign error.
Moreover, I had a different kinematic with respect to the preprint with
different signs for the angles z_gamma and z_w.
That lead to a different sign in six terms of  angle-dependent
observables. It does not change anything in the other observables.
Especially, does it not change the asymmetry since I implemented
the asymmetry already with the corrected set of kinematics.