[python] Change analysis file format
Since we now support correlated multivariate prior components, the use
of ``parameters`` as a key in the serialization of a prior component is
not apt anymore. This has been changed to ``descriptions``, with
backward compatibility for older files.

Error messages have been adjusted to reflect wording of the
documentation and the draft paper for v1.1.