[eos] Add a create-mask task
Allows the creation of a mask on a set of ImportanceSamples, based upon
the sign of one or more observable(s), and when using multiple observable
the individual masks can be combined using either 'and' or 'or' logic.
Masks can be used in the find-mode, corner-plot, and predict-observables
tasks after creation.