[rare-b-decays] update Bs -> phi mumu based on arXiv 1502.05509
BsToPhiDilepton: Add two public functions to retrieve the integrated and differential angular coefficients.
Adjusted the prefactors for the J_i to use the convention from Virto et al in 1502.05509.
This also means that some pre-factors in functions change slightly.

BsToPhiDileptonAndConjugate:
Add new properties m_y, m_x, m_gamma_s corresponding to the DeltaGamma/Gamma_s, DeltaM/Gamma_s and Gamma_s.
Add new struct: AngularsCoefficients mirrors AngularhCoefficients.
Add calculation of the s_i coefficients.

Modify AngularhCoefficients to use the same prefactors as in Virto et al 1502.05509.
Fix a typo in h_1c.
Add functions for the integrated and differential angular coefficients S, A, K, W, H, and Z.
The K and W correspond to the "S" and "A" in the Kst case and used to be called S and A for the phi case as well.
The new S and A are based on Virto et al 1502.05509 and reflect on the time-integration needed to be performed which introduces factors of y / x and h / s.
The definition of the functions is handled using macros to avoid typos and repetetive typing.