[rare-b-decays] add new Bs -> phi observables.
B_s->phi_ll::Gamma uses the implementation of BsToPhiDileptonAndConjugate reflecting on Virto et al in 1502.05509.
Add all K,W,H,Z,S, and A observables in integrated and differential as well as "theory" and "LHCb" convention.
Change between "theory" and "LHCb" is a minus sign for coefficients 4, 6s, 6c, 7, and 9, see Figure 4 in 1506.03970.